vscode: gracefully handle cancellation errors
